Many small corrections to PersistentDataStructures transcript
I know that a lot of the changes here are simply expanding English
contractions, e.g. can't -> cannot, which is a personal preference of
mine when reading text vs. listening to talks.

But there are also many small corrections to other words Rich said
during the talk, from carefully listening to the talk twice through
and correcting differences as I found them.